requirements.txt
streamlit>=1.32.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
pytesseract>=0.3.10
spacy>=3.7.0
rapidfuzz>=3.6.0
scikit-learn>=1.4.0
sentence-transformers>=2.6.0
plotly>=5.19.0
google-generativeai>=0.5.0
fpdf2>=2.7.0
chardet>=5.2.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
